Split workflow config modules
